cleanup old experiment stuff
